RavenDB-6609 - moved index creation/deletion to raft - fixing tests

+++ src/Raven.Server/Documents/Indexes/Errors/FaultyInMemoryIndex.cs
@@ -23,7 +23,7 @@ namespace Raven.Server.Documents.Indexes.Errors
     {
         private readonly Exception _e;
 
-        public FaultyInMemoryIndex(Exception e, int etag, string name, IndexingConfiguration configuration)
+        public FaultyInMemoryIndex(Exception e, long etag, string name, IndexingConfiguration configuration)
             : base(etag, IndexType.Faulty, new FaultyIndexDefinition(name ?? $"Faulty/Indexes/{etag}", new HashSet<string> { "@FaultyIndexes" },
                    IndexLockMode.Unlock, IndexPriority.Normal, new IndexField[0]))
         {
